+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Changes relevant to the users of python-tcod are documented here.
 This project adheres to [Semantic Versioning](https://semver.org/) since version `2.0.0`.
 
 ## [Unreleased]
+### Added
+- `BasicMixer` and `Channel` classes added to `tcod.sdl.audio`.  These handle simple audio mixing.
+- `AudioDevice.convert` added to handle simple conversions to the active devices format.
+- `tcod.sdl.audio.convert_audio` added to handle any other conversions needed.
 
 ## [13.5.0] - 2022-02-11
 ### Added